requirements.txt
fastapi
uvicorn
passlib
pymongo==4.3.3
pymongo[zstd]
pymongo[srv]
python-jose
python-multipart